* commit '07f10d2a94e7342413fa7762b6ce6b101b93d130': (86 commits)
  Move optimizer to create_optimizer (#2851)
  support reward_model (#2849)
  1. fix hub ignore-pattern (#2848)
  Fix bugs (#2838)
  Update base_to_chat shell (#2833)
  Update padding side (#2832)
  Fix glm4v suffix (#2829)
  add 'right' option for 'truncation_strategy' (#2754)
  update docs (specific model arguments) (#2822)
  fix enable_cache (#2813)
  fix citest (#2812)
  support ZhipuAI/cogagent-9b-20241220 (#2810)
  fix swift deploy log error (repeat log) (#2808)
  fix glm4v (#2806)
  update base_model deploy example (#2803)
  fix world_size (#2801)
  fix (#2800)
  support swift app (#2792)
  fix some web-ui bugs (#2794)
  fix stream infer (#2793)
  ...

# Conflicts:
#	examples/train/multi-gpu/ddp/train.sh
#	swift/llm/__init__.py
#	swift/llm/argument/rlhf_args.py
#	swift/llm/template/base.py
#	swift/llm/template/template_inputs.py
#	swift/llm/template/utils.py
#	swift/llm/train/tuner.py
#	swift/trainers/mixin.py

+++ swift/ui/llm_infer/runtime.py
@@ -95,7 +95,7 @@ class Runtime(BaseUI):
         with gr.Accordion(elem_id='runtime_tab', open=False, visible=True):
             with gr.Blocks():
                 with gr.Row():
-                    gr.Dropdown(elem_id='running_tasks', scale=10)
+                    gr.Dropdown(elem_id='running_tasks', scale=10, allow_custom_value=True)
                     gr.Button(elem_id='refresh_tasks', scale=1, variant='primary')
                     gr.Button(elem_id='show_log', scale=1, variant='primary')
                     gr.Button(elem_id='stop_show_log', scale=1)
@@ -131,7 +131,7 @@ class Runtime(BaseUI):
         latest_data = ''
         lines = collections.deque(maxlen=int(os.environ.get('MAX_LOG_LINES', 50)))
         try:
-            with open(log_file, 'r') as input:
+            with open(log_file, 'r', encoding='utf-8') as input:
                 input.seek(offset)
                 fail_cnt = 0
                 while True:
@@ -233,13 +233,14 @@ class Runtime(BaseUI):
 
     @classmethod
     def kill_task(cls, task):
-        pid, all_args = cls.parse_info_from_cmdline(task)
-        log_file = all_args['log_file']
-        if sys.platform == 'win32':
-            os.system(f'taskkill /f /t /pid "{pid}"')
-        else:
-            os.system(f'pkill -9 -f {log_file}')
-        time.sleep(1)
+        if task:
+            pid, all_args = cls.parse_info_from_cmdline(task)
+            log_file = all_args['log_file']
+            if sys.platform == 'win32':
+                os.system(f'taskkill /f /t /pid "{pid}"')
+            else:
+                os.system(f'pkill -9 -f {log_file}')
+            time.sleep(1)
         return [cls.refresh_tasks()] + [gr.update(value=None)]
 
     @classmethod
@@ -266,9 +267,4 @@ class Runtime(BaseUI):
                 ret.append(gr.update(value=arg))
             else:
                 ret.append(gr.update())
-        train_type = None
-        if is_custom_path:
-            with open(os.path.join(all_args['ckpt_dir'], 'args.json'), 'r') as f:
-                _json = json.load(f)
-                train_type = _json.get('train_type')
-        return ret + [gr.update(value=None), [all_args.get('model_type'), all_args.get('template_type'), train_type]]
+        return ret + [gr.update(value=None)]
